requirement.txt
streamlit
pandas
google.generativeai
io
re